backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
litellm==1.55.10
python-multipart==0.0.20
websockets==14.1
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
